+++ _test/test/goldens/generated_build_script.dart
@@ -1,4 +1,5 @@
 // ignore_for_file: directives_ordering
+// ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'package:build_runner_core/build_runner_core.dart' as _i1;
 import 'package:build_test/builder.dart' as _i2;
 import 'package:build_config/build_config.dart' as _i3;